scripts/step03_explore_examples/utils/pyproject.toml
[tool.poetry]
name = "utils"
version = "0.1.4"
description = "Utils for exploring examples"
authors = ["romaniam"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
hdbcli = "^2.21.26"
langchain = "^0.2.6"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
